feat: add bad request error

+++ src/resources/boadrs/boards.router.js
@@ -2,6 +2,7 @@ const router = require('express').Router();
 const boardsService = require('./boards.service');
 const tasksService = require('../tasks/tasks.service');
 const { errorHandler } = require('../../common/error-handler');
+const { BAD_REQUEST, NOT_FOUND } = require('http-status-codes');
 
 router
   .route('/')
@@ -10,6 +11,10 @@ router
     res.json(boards);
   })
   .post(async (req, res) => {
+    if (!req.body.title || !req.body.columns) {
+      errorHandler(res, BAD_REQUEST, 'Bad request');
+      return;
+    }
     const board = await boardsService.createBoard(req.body);
     res.json(board);
   });
@@ -21,12 +26,16 @@ router
     if (board) {
       res.json(board);
     } else {
-      errorHandler(res, 'no boards with this ID');
+      errorHandler(res, NOT_FOUND, 'no boards with this ID');
     }
   })
   .put(async (req, res) => {
     const user = await boardsService.getBoard(req.params.id);
-    res.json(boardsService.updateBoard(user, req.body));
+    if (user) {
+      res.json(boardsService.updateBoard(user, req.body));
+    } else {
+      errorHandler(res, NOT_FOUND, 'no boards with this ID');
+    }
   })
   .delete(async (req, res) => {
     const board = await boardsService.getBoard(req.params.id);
@@ -34,7 +43,7 @@ router
       boardsService.deleteBoard(board);
       res.json('The user has been deleted');
     } else {
-      errorHandler(res, 'no board with this id');
+      errorHandler(res, NOT_FOUND, 'no board with this ID');
     }
   });
 
@@ -45,6 +54,16 @@ router
     res.json(tasks);
   })
   .post(async (req, res) => {
+    if (
+      !req.body.title ||
+      req.body.order === undefined ||
+      !req.body.description ||
+      req.body.userId === undefined ||
+      req.body.boardId === undefined
+    ) {
+      errorHandler(res, BAD_REQUEST, 'Bad request');
+      return;
+    }
     const task = await tasksService.createTask(req.params.id, req.body);
     res.json(task);
   });
@@ -56,7 +75,7 @@ router
     if (task) {
       res.json(task);
     } else {
-      errorHandler(res, 'no tasks');
+      errorHandler(res, NOT_FOUND, 'no tasks with this ID');
     }
   })
   .put(async (req, res) => {
@@ -65,8 +84,12 @@ router
   })
   .delete(async (req, res) => {
     const task = await tasksService.getTaskById(req.params.taskId);
-    tasksService.deleteTask(task);
-    res.json('The task has been deleted');
+    if (task) {
+      tasksService.deleteTask(task);
+      res.json('The task has been deleted');
+    } else {
+      errorHandler(res, NOT_FOUND, 'no tasks with this ID');
+    }
   });
 
 module.exports = router;
